backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
pylint==3.0.3
flake8==6.1.0
black==23.11.0
bandit==1.7.5
radon==6.0.1
vulture==2.10
mypy==1.7.1
astroid==3.0.1
tree-sitter==0.20.4
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.1
tree-sitter-java==0.23.1
requests==2.31.0
Pillow==10.1.0
gunicorn==21.2.0
# Control flow analysis dependencies
graphviz==0.20.1
pydot==2.0.0
networkx==3.2.1
numpy==1.26.4
nest_asyncio==1.6.0
plotly==5.18.0
pyvis==0.3.2
pandas==2.2.0
werkzeug==3.0.1
